Added user tests

+++ lib/udia/web/controllers/user_controller.ex
@@ -13,12 +13,19 @@ defmodule Udia.Web.UserController do
     render(conn, "index.json", users: users)
   end
 
-  def create(conn, %{"user" => user_params}) do
-    with {:ok, %User{} = user} <- Accounts.create_user(user_params) do
-      conn
-      |> put_status(:created)
-      |> put_resp_header("location", user_path(conn, :show, user))
-      |> render("show.json", user: user)
+  def create(conn, user_params) do
+    case Accounts.create_user(user_params) do
+      {:ok, user} ->
+        new_conn = Guardian.Plug.api_sign_in(conn, user, :access)
+        jwt = Guardian.Plug.current_token(new_conn)
+
+        new_conn
+        |> put_status(:created)
+        |> render(Udia.Web.SessionView, "show.json", user: user, jwt: jwt)
+      {:error, changeset} ->
+        conn
+        |> put_status(:unprocessable_entity)
+        |> render(Udia.Web.ChangesetView, "error.json", changeset: changeset)
     end
   end
 
@@ -29,16 +36,40 @@ defmodule Udia.Web.UserController do
 
   def update(conn, %{"id" => id, "user" => user_params}) do
     user = Accounts.get_user!(id)
+    cur_user = Guardian.Plug.current_resource(conn)
 
-    with {:ok, %User{} = user} <- Accounts.update_user(user, user_params) do
-      render(conn, "show.json", user: user)
+    if cur_user.id != user.id do
+      conn
+      |> put_status(:forbidden)
+      |> render(Udia.Web.SessionView, "forbidden.json", error: "Invalid user")
+    else
+      case Accounts.update_user(user, user_params) do
+        {:ok, user} ->
+          new_conn = Guardian.Plug.api_sign_in(conn, user, :access)
+          jwt = Guardian.Plug.current_token(new_conn)
+
+          new_conn
+          |> put_status(:created)
+          |> render(Udia.Web.SessionView, "show.json", user: user, jwt: jwt)
+        {:error, changeset} ->
+          conn
+          |> put_status(:unprocessable_entity)
+          |> render(Udia.Web.ChangesetView, "error.json", changeset: changeset)
+      end
     end
   end
 
   def delete(conn, %{"id" => id}) do
     user = Accounts.get_user!(id)
-    with {:ok, %User{}} <- Accounts.delete_user(user) do
-      send_resp(conn, :no_content, "")
+    cur_user = Guardian.Plug.current_resource(conn)
+    if cur_user.id != user.id do
+      conn
+      |> put_status(:forbidden)
+      |> render(Udia.Web.SessionView, "forbidden.json", error: "Invalid user")
+    else
+      with {:ok, %User{}} <- Accounts.delete_user(user) do
+        send_resp(conn, :no_content, "")
+      end
     end
   end
 end
